factorial return value changed to double

+++ src/calc.cpp
@@ -77,14 +77,14 @@ double Math::powerOf(double operand, int exponent) {
 	return 0.0;
 }
 
-int Math::factorial(int operand) {
-	int result = 1;
+double Math::factorial(int operand) {
+	double result = 1;
 	try {
 		if (operand < 0) {
 			throw std::range_error("Number cannot be negative!");
 		} else {
 			for (int i = 1; i <= operand; i++) {
-				result = multiplication(result, i);
+				result = multiplication(result, double(i));
 			}
 		}
 	} catch (std::range_error &e) {
